imported EXPECT_EQUAL and updated toolkit from feature/ServerManager

+++ Toolkit/Arrays/DynArray/_Tests/test2.cpp
@@ -150,7 +150,7 @@ int TestPart2(int testNumber)
 
         EXPECT_EQUAL(std.size(), assign.size(), "::move failed, size mismatch");
 
-        EXPECT_EQUAL(array.size(), (size_t)0, "::move failed, source array not empty");
+        EXPECT_EQUAL(array.size(), 0, "::move failed, source array not empty");
 
         array.push_back(1);
         array.push_back(2);
@@ -159,7 +159,7 @@ int TestPart2(int testNumber)
 
         array.clear();
 
-        EXPECT_EQUAL(array.size(), (size_t)0, "::clear failed, array not empty");
+        EXPECT_EQUAL(array.size(), 0, "::clear failed, array not empty");
 
 		std::cout << "	PASSED" << std::endl;
 	}
